% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_GIDs.R
\name{Check_GIDs}
\alias{Check_GIDs}
\title{Check GIDs}
\usage{
Check_GIDs(GIDs, crop = F)
}
\arguments{
\item{list_of_GIDs}{numeric - an integer of GIDs}
}
\value{
Retunrs a message if all GIDs are valid or which are invalid. If 'crop=TRUE' returns a vector of all valid input GIDs
}
\description{
Check if input GIDs are valid
}
\examples{
# get vector of GIDs
gid <- c(522647, 522710, 1)
# check
Check_GIDs(gid)
# crop invalid GIDS
gid_new <- Check_GIDs(gid,crop=T)
gid_new
}
\author{
Andreas Schönberg
}
