% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMatDR-Package.R
\docType{package}
\name{DistMatDR - package}
\alias{DistMatDR - package}
\title{Distance Matrix of traveltime by foot for the 'German Reich'}
\description{
The \code{DistMatDR} package is used to get the traveltime on foot between desired GIDs in the 'German Reich'.
}
\details{
DistMatDR Package
}
\author{
Andreas Schönberg
}
\keyword{package}
