% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DistMat.R
\name{get_DistMat}
\alias{get_DistMat}
\title{Get DistMat}
\usage{
get_DistMat(GIDs)
}
\arguments{
\item{GIDs}{numeric - a vector of GIDs}
}
\value{
Returns
}
\description{
Check if input GIDs are valid
}
\examples{
# load data
data("valid_GIDs")

# get vector of random GIDs
gids <- sample(valid_GIDs,102)
res <- get_DistMat(gids)
}
\author{
Andreas Schönberg
}
