% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knn.R
\name{get_knn}
\alias{get_knn}
\title{Get knn and traveltime for GID}
\usage{
get_knn(GID, k)
}
\arguments{
\item{GID}{numeric - a vector of GIDs}

\item{GIDs_to}{numeric - a vector of GIDs}
}
\value{
Returns
}
\description{
sdf
}
\examples{
# get the 5 nearest neighbours for GID 112019
res <-get_knn(112019,5)
res
}
\author{
Andreas Schönberg
}
