% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traveltime_from_to.R
\name{get_traveltime_from_to}
\alias{get_traveltime_from_to}
\title{Get traveltime from source GIDs to target Gids}
\usage{
get_traveltime_from_to(GIDs_from, GIDs_to)
}
\arguments{
\item{GIDs_from}{numeric - a vector of GIDs}

\item{GIDs_to}{numeric - a vector of GIDs}
}
\value{
Returns
}
\description{
sdf
}
\examples{
# load data
data("valid_GIDs")
gids_from <-sample(valid_GIDs,20)
gids_to <-sample(valid_GIDs,5)

res <-get_traveltime_from_to(GIDs_from = gids_from,GIDs_to = gids_to)
res
}
\author{
Andreas Schönberg
}
