% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFun.R
\name{TestFun}
\alias{TestFun}
\title{TestFun}
\usage{
TestFun(x)
}
\arguments{
\item{x}{numeric - a test value}
}
\value{
if all wokrs the function returns a message
}
\description{
test the package
}
\examples{
TestFun(2019)
}
\author{
Andreas Schönberg
}
