% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traveltime.R
\name{get_traveltime}
\alias{get_traveltime}
\alias{Check_GIDs}
\title{get_traveltime}
\usage{
get_traveltime(GIDs)
}
\arguments{
\item{GIDs}{numeric - an integer of GIDs}
}
\value{
Returns the Distacne Matrix (traveltime on foot) for the given GIDs for both directions.
}
\description{
Extract the traveltime distance matirx for a given set of GIDs
}
\examples{

# get vector of GIDs

gid <- sample(valid_GIDs,100)
# check
df<-get_traveltime(gid)
}
\author{
Andreas Schönberg
}
