% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traveltime.R
\name{get_traveltime}
\alias{get_traveltime}
\title{get_traveltime}
\usage{
get_traveltime(GIDs)
}
\arguments{
\item{GIDs}{numeric - an integer of GIDs}
}
\value{
Returns the Distacne Matrix (traveltime on foot) for the given GIDs for both directions.
}
\description{
Extract the traveltime distance matirx for a given set of GIDs
}
\examples{

# load data
data("valid_GIDs")

# get vector of random GIDs
gids <- sample(valid_GIDs,10)
# get traveltime distanc matrix
res<-get_traveltime(gids)
res
}
\author{
Andreas Schönberg
}
